+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+* Release BigQuery sink with at-least-once support.
+* More details to be added soon!
+
 ## 0.1.0-preview - 2023-12-14
 
 * Initial release with BQ source support